configure lsp and ipython
After some deliberation, I decided that ipython worked better for my use
cases than python3 as an interpreter. I also set up basedpyright as an
lsp for python files. This does require the use of a pyrightconfig.json
file in the root of the project directory to fully set it up. Working on
figuring out how to configure basedpyright globally.